fix data reset error on in progress interview session
