+++ public/components/windowManager.js
@@ -1,33 +1,12 @@
 const { BrowserWindow, Menu, shell } = require('electron');
-const url = require('url');
-const path = require('path');
 const mainMenu = require('./mainMenu');
-const log = require('./log');
+const appUrl = require('./appURL');
 
 const isMacOS = () => process.platform === 'darwin';
 
 const titlebarParameters = isMacOS() ? { titleBarStyle: 'hidden', frame: false } : {};
 
 let window;
-
-const appUrl = (function getAppUrl() {
-  if (process.env.NODE_ENV === 'development' && process.env.NC_DEVSERVER_FILE) {
-    // This method is more robust than Architect & Server to support multiple platforms & devices
-    // NC_DEVSERVER_FILE contains the URL of a running webpack-dev-server, relative to app root
-    try {
-      const relativePath = path.join(__dirname, '..', '..', process.env.NC_DEVSERVER_FILE);
-      return require('fs').readFileSync(relativePath, 'utf-8'); // eslint-disable-line global-require
-    } catch (err) {
-      log.warn('Error loading dev server config -', err.message);
-      log.warn('Are you running dev server?');
-      log.warn('Continuing with index.html');
-    }
-  }
-  return url.format({
-    pathname: path.join(__dirname, '..', 'index.html'),
-    protocol: 'file:',
-  });
-}());
 
 function setApplicationMenu(appWindow) {
   const appMenu = Menu.buildFromTemplate(mainMenu(appWindow));
